7.3 > パスを`pub`キーワードで公開する
