--- the-book/restaurant/src/front_of_house.rs
@@ -1,1 +1,0 @@
-pub mod hosting;
