--- the-book/restaurant/src/front_of_house/hosting.rs
@@ -1,11 +1,0 @@
-pub fn add_to_waitlist() {}
-
-fn seat_at_table() {}
-
-mod serving {
-    fn take_order() {}
-
-    fn serve_order() {}
-
-    fn take_payment() {}
-}
