+++ the-book/restaurant/src/lib.rs
@@ -1,50 +1,23 @@
-mod front_of_house;
+mod front_of_house {
+    pub mod hosting {
+        pub fn add_to_waitlist() {}
 
-use crate::front_of_house::hosting;
-
-mod back_of_house {
-    pub enum Appetizer {
-        Soup,
-        Salad,
+        fn seat_at_table() {}
     }
 
-    pub struct Breakfast {
-        pub toast: String,
-        seasonal_fruit: String,
-    }
+    mod serving {
+        fn take_order() {}
 
-    impl Breakfast {
-        pub fn summer(toast: &str) -> Breakfast {
-            Breakfast {
-                toast: String::from(toast),
-                seasonal_fruit: String::from("peaches"),
-            }
-        }
+        fn serve_order() {}
+
+        fn take_payment() {}
     }
 }
-
-// relative path
-// use self::front_of_house::hosting;
 
 pub fn eat_at_restaurant() {
     // Absolute path
-    hosting::add_to_waitlist();
+    crate::front_of_house::hosting::add_to_waitlist();
 
     // Relative path
     front_of_house::hosting::add_to_waitlist();
-
-    // Order a breakfast in the summer with Rye toast
-    let mut meal = back_of_house::Breakfast::summer("Rye");
-    // Change our mind about what bread we'd like
-    meal.toast = String::from("Wheat");
-    println!("I'd like {} toast please", meal.toast);
-
-    // The next line won't compile if we uncomment it; we're not allowed
-    // to see or modify the seasonal fruit that comes with the meal
-    // meal.seasonal_fruit = String::from("blueberries");
-
-    // Access to enum
-    let order1 = back_of_house::Appetizer::Soup;
-    let order2 = back_of_house::Appetizer::Salad;
 }
-
